Stop dropping the CONTENT_TYPE header when forwarding requests
RequestForwarder copies headers by looking for HTTP_ prefixes in
the Rack-parsed environment given to each request, and explicitly
hacking the CONTENT_LENGTH (which is not prefixed) to align with
the transfer encoding of the body.

However, CONTENT_TYPE is also a special case (see
https://github.com/rack/rack/issues/1311) and must be explicitly
checked for. As a result, content-store-proxy was not passing the
content type through, and this caused issues for POST requests.
In the absence of an explicit JSON content-type, Rails was
treating the JSON in the body as form-encoded, and this led to
problems with URLs embedded in the body content - ampersands were
getting Unicode-escaped to \u0026, and the whole request was being
dropped by content-store with a HTTP 400 Bad Request.

This PR fixes the issue by explicitly checking for a given
CONTENT_TYPE and passing it through.