Handle all errors on the secondary_response more gracefully
To stop errors on the secondary at the time the Mongo-to-Postgresql
cron job completes propagating through to the frontends and users,
we need to handle errors on the secondary response more gracefully.

It now catches ConnectionFailed as well as TimeoutError, and logs
the message & type in an error: key in the secondary response
JSON packet.